+++ src/main/kotlin/com/github/sleepypanda/feesh/utils/data/PersonalBestData.kt
@@ -11,6 +11,9 @@ data class PersonalBestData(
     val sharksCaught: PersonalBestEntry = PersonalBestEntry(),
     val greatWhiteSharksCaught: PersonalBestEntry = PersonalBestEntry(),
     val doubleHookStreak: PersonalBestEntry = PersonalBestEntry(),
+    val treasureCatchesStreak: PersonalBestEntry = PersonalBestEntry(),
+    val greatTreasuresStreak: PersonalBestEntry = PersonalBestEntry(),
+    val outstandingTreasuresStreak: PersonalBestEntry = PersonalBestEntry(),
     val totalMobyDucksConsumed: PersonalBestEntry = PersonalBestEntry(),
 )